+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Version 2.2.1 (Under development)
 
+### App Center Push
+
+* **[Fix]** Fix confusing information log about the availability of the Firebase SDK.
+
 ### App Center Auth
 
 * **[Fix]** Redirect URIs are now hidden in logs.
+
 ___
 
 ## Version 2.2.0